add : toolsStreamProvider
