--- lib/widget/tools/lesson_start_slide.dart
@@ -1,86 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_swipe_button/flutter_swipe_button.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-
-import '../../data/firebase/during_stream.dart';
-import '../../data/firebase/lesson_stream.dart';
-import '../../data/record/record.dart';
-import '../utils/sakura_progress_indicator.dart';
-
-class LessonStartSlide extends ConsumerWidget {
-  const LessonStartSlide({
-    super.key,
-    this.width = 400,
-    this.height = 60,
-  });
-
-  final double width, height;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final during = ref.watch(duringStreamProvider);
-    final currentRoom = ref.watch(currentRoomProvider);
-    final lesson = ref.watch(currentLessonStreamProvider);
-    final recorder = ref.watch(streamRecorderProvider);
-    final recorderNot = ref.read(streamRecorderProvider.notifier);
-
-    return during.when(
-      data: (snapshot) {
-        switch (duringLesson(
-          snapshotData: snapshot.docs,
-          roomNumber: currentRoom.roomNumber,
-          subject: currentRoom.subject,
-          count: lesson!.data().count,
-        )) {
-          case true:
-            return SwipeButton.expand(
-              width: width,
-              height: height,
-              child: Text("授業終了"),
-              onSwipeEnd: () async {
-                if (lesson.data().state != "break") {
-                  await cancelLessonToDuring(
-                    teacher: currentRoom.teacher,
-                    currentLesson: lesson.data(),
-                  );
-                } else {
-                  await finishLessonToDuring(
-                    teacher: currentRoom.teacher,
-                    currentLesson: lesson.data(),
-                  );
-                }
-                if (recorder.isRecording) {
-                  await recorderNot.stop();
-                }
-              },
-            );
-          case false:
-            return Text("他のコマで授業中です");
-          default:
-            return SwipeButton.expand(
-              width: width,
-              height: height,
-              child: Text("授業開始"),
-              onSwipeEnd: () async {
-                await addLessonToDuring(
-                  roomNumber: currentRoom.roomNumber,
-                  subject: currentRoom.subject,
-                  count: lesson.data().count,
-                  teacher: currentRoom.teacher,
-                  currentLesson: lesson.data(),
-                );
-              },
-            );
-        }
-      },
-      // エラー時の表示
-      error: (_, __) => const Center(
-        child: Text("読み込み失敗"),
-      ),
-      // 読込中の表示
-      loading: () => const Center(
-        child: SakuraProgressIndicator(),
-      ),
-    );
-  }
-}
